dev_requirements.txt
black==19.10b0
falcon-crossorigin==0.1.2
fastjsonschema==2.14.4
mock==4.0.2
pre-commit==2.3.0
pytest==5.4.2

-r requirements.txt
